Have created "svg-icon" CSS class.
This is used to style the positioning and size of the SVG icons that will be in the navigation bar for the web app. Have also added hr border element styling for break line that will be in all web pages just under the nav bar. Has been added to the index & help web pages so far.