Have made changes to global nav bar.
To start with on the landing page (index.ejs), I have added the title card that has been placed above the global nav bar which most of the web pages will contain within the web app. This is where the "TASKACARD" title is shown. The nav bar buttons have been adjusted and colouring variables have been added and used that are stored within the :root{} class in the global.css file.