Have created and linked official stylesheet for help page.
Have created specific classes such as the "drop-down" button class that is located within in a wide-card. Each of these buttons will contain a frequently asked question for the help page and when clicked on will act as a collapsible drop down revealing the answer which will be stored in the "answer-content" class.