Have coded the JS functionality for button collapsible.
The frequently asked questions on this web page now has the collapsible feature whereby if a question button is clicked on, the answer drop down will be activated. The "answer-content" CSS class containing the answer text will be toggled active in the list of next sibling elements after the "arrow-icon" class using the variable named "content" in the if statement.

Have also code commented accordingly.