Have coded the Kanban data storage configuration.
Have coded the following functions within the KanbanAPI js class: "getItems, insertItem, updateItem & deleteItem". These four key functions deal with the CRUD functionality of task card data being modified on the kanban-board web page. The "updateItem" function also contains the updating of the position for the task card when it is dropped into another Kanban column on the web page. This uses the items.splice method which can move the array of data content within that task card to another column.

The "read" function gets the data items from the data key "board-data" and returns each kanban column as a JSON array. Each column/array ID is numbered from 1-3.

The "save" function sets the "board-data" storage key to it's updated data when text content within a task card is changed. This uses the JSON.stringify for data.