Have created the "column" class within the column.js file.
This class is used for dynamically generating the HTML content for each of the three Kanban columns on the board. This uses the JS "elements" objects. A root is used to query select one of the div class name elements from the "kanban" HTML template so content can be generated from the JS. The Contextual Fragment acts like a virtual DOM tree.