Have created the "kanban" class within the kanban.js file.
The "column" class has been imported from the "column.js" file as the HTML elements that will be dynamically generated will be placed into each of the three static array ID's labelled with "Not Started, In Progress & Complete". These are what link to represent the three columns on the Kanban board.