Have started to create "Card" class within card.js file.
This class is what is responsible for inputting content into a single task card element. The "onBlur" object is what persists and saves text content for a task card if it has been modified and is clicked out of. If the web page is refreshed, the content will remain and persist as it is linked to the local storage data key. This makes use of the ".taskcard-input" class using querySelector.