Have created the "DropZone" class within the "DropZone.js" file.
This class is responsible for dealing with the functionality for the drop zone that appears when a task card is being dragged and dropped from it's current Kanban column to a new one. The JS "dragover", "dragleave" & "drop" event listener functions trigger this. The position index of a dropped task card is saved and updated once it has been dropped to a new state of position.