DropZone class has been imported into the "Card" class.
The const "bottomDropZone" is the value of the DropZone that calls the "createDropZone();" function that is located in the DropZone class. Here the bottom dropzone is configured as the bottom drop zone will always appear underneath a single task card in a column when a dropping is being toggled over the drop area.