DropZone class has been imported into the "Column" class.
The const "topDropZone" is the value of the DropZone that calls the "createDropZone();" function that is located in the DropZone class. Here the top dropzone is configured as the top drop zone will always appear above a single task card in a column when a dropping is being toggled over the drop area. Kanban-board page functionality is now complete.