Have started to make "main.hbs" file.
This is the handlebars file that is stored within the newly created "layouts" folder which is a sub folder of the "views" folder. This file will act as the main template body parser for the relevant hbs web pages that will be created for the personnel management system feature. Then curly braces have been set within the HTML body tag.