Have included Nav bar for Personnel Management Table.
Have included a secondary nav bar as opposed to the global nav bar on all of the other web pages. This is the Personnel Management Table nav bar that will be present along all of the four web pages relating to this feature of the web app. This includes the home-table web page and the add, view & edit personnel record pages.

Have linked the bootstraps stylesheet to "main.hbs" and the Bootstrap icons CDN stylesheet link. The nav bar classes are from and make use of the bootstrap styling.