Have started to code SQL database queries (API endpoints).
Within the index.js file, two API endpoint functions for the express API have been coded so far.

These include the app.get function that contains the MySQL query for selecting all records from the "personnel" table in the schema where only the "active" personnel users are to be displayed. The "rows" object is used as a parameter to store and pass through the database query. This rows object is then declared within the "home-table.hbs" web page using the  {{#each rows}} {{/each}} tag setup. In-between these tags is the HTML tables <tr> and <td> tags so a records of data containing the correct column values for data can be rendered and pushed through from the database table.

The second API endpoint function that has been coded so far is the app.post function that contains the MySQL query for searching for specific users (records of data) from the table when using the search input field. So far this has only been enabled by the these two table attribute names which are first_name & last_name. These two fields used '%' + searchRecord + '%'   where searchRecord is the name of the JS variable used to store the req.body.search. ".search" is the class name in the HTML for the input form where string text is entered and the API can communicate with the database to pull those search results form the table.