Personnel table search field input box: configuration.
Here within the "main.hbs" render layout, the form class is being given a POST method in the HTML so it can work in conjunction with the POST API endpoint that contains the SQL query to search for personnel users. The name "search" is the key name being used to equal as the JS variable for that SQL query which is called "searchRecord".