Have started to create the "add-personnel.hbs" web page.
This is the second handlebars web page view that is being created for the personnel table feature. The web page contains a form layout for a new personnel details record to be created for the system.

Have added the bootstrap breadcrumbs inner navigation bar feature that can navigate the user back to the home table page when at another page in the table system.

Have also added the bootstrap form floating feature within the form class in the HTML. So far it contains an email and password input box form but will be modified. The main form that wraps all of this content has also been given a method of "POST" and no validation.