Have coded the API end point for Adding new record.
This POST function includes the MySQL query for adding a new personnel record into the personnel table. This is done through the input box forms and is rendered on the "add-personnel.hbs" page. The INSERT INTO & SET sql statements are used to insert the relevant data properties into their fields for a personnel record. The "?" question mark symbol represents an unnamed parameter which can be filled in by the API running the query to avoid SQL injection. This was also used in the "View personnel records" query before which is used to display the whole table.