Added corresponding API endpoint content to "add-personnel.hbs" file.
This includes the HTML form class that acts as a POST Method and is actioned with the "/add-personnel" API end point route. This is the form that works in conjunction to the "add new personnel record" SQL query. Each input form area contains a label representing the input fields for "First Name, Last Name, Email, Telephone, Comments". These also contain a "name" value within each form input type. This "name" value is what corresponds with the variable parameter names in the POST API function for adding a new personnel record just before the SQL query.

The {{alert}} message has also been placed into the {{#if alert}} curly braces which prompts the user that they have added a new personnel record successfully. This is linked to the "alert" object and it's message that is rendered within the POST function for this web page.