Have created the "personnel-form" hbs partial file.
Have placed this within a "partials" folder in the project work space. This file stores the HTML for the personnel input boxes that are used in the add, edit view and edit update options for a personnel record in the system table. This saves repeating the HTML in each of the web page templates within the main form tag.