Added corresponding API endpoint content to "edit-personnel.hbs" file.
This includes the HTML form class that acts as a POST Method and is actioned with the "/edit-personnel" API end point route. This is the form that works in conjunction to the "edit (view) personnel record" SQL query option. Have included the "personnel-form" partial file link in-between the form tags. Have also included the breadcrumbs inner navigation bar.

Have coded the SQL query within the app.get API end point method that is responsible for selecting a personnel record id for viewing the data fields within the edit view. This done with "SELECT * FROM personnel WHERE id = ?" and uses the [req.params.id].

The values within the "personnel-from.hbs" partial file for each input field represent the parameter field names from the table. For example first_name is equal to "first_name" as it is within the MySQL database.