(Have began coding Login feature) Have coded the "sign-up" and "login" method routes.
Both the API post functions contain the SQL queries for their respective functionalities.

The "/sign-up" query contains the INSERT statement for the username, email & password input fields on the "user-sign-up.ejs" page. These fields get filled out and represent a new user record that has been placed into the "users" table within the db schema.

The "/login" query contains the SELECT statement for the username & password where if the password === password (from the database for that specific username) and is correct the "index.ejs" page will be rendered bringing the logged in user to the landing page. Web session tokens now need to be dealt with. The corresponding POST form methods also contain the correct "name" values within the "login" and "user-sign-up" page.