Have enabled user session tokens for user login feature.
With requiring the "crypto" dependency, this is used to encrypt and generate a session token that is assigned to an exiting userID when they log in, thus assigning that userID  to a sessionID with that token in the "user_login" db table.

Within the "/login" API route method, the variable "token" has been assigned to the crypto.randomBytes. This is used to generate the session token string for a logged in sessionID. Then the MySQL query is used to insert the 'userID' & it's 'login_token' into a session row in the 'user_login' table. If the username and password match, as well as the password matching an existing password, the user will be taken to the landing page as the "index.ejs" page will be rendered. This would take place after a user logging in successfully on the "login.ejs" web page.

To start with, the "home-table.hbs" web page within the app.get("/home-table") route has been given and required the userID & token value variables. The MySQL query select statement grabs all of the rows from the 'user_login' table to see if an existing login token matches. If all matches for a currently logged in user session, the user will be able to view the "home-table.hbs" page. This needs to be configured for all web pages.

Have also code commented the above content so far with short detailed descriptions.