Have included alert message when personnel record is deleted.
Have created and included the "removedUser " object that represents the encodeURIComponent removal for that data row. This has been configured in the sql query for deleting a personnel record and the view all query. When the removedUser object spawns for any row ID being deleted in the table, an alert message will inform the user of deleting a personnel record.

The {{#if removedUser}} in "home-table.hbs" is used for the wrapping the alert class box message.