Have included the [contenteditable] CSS class for kanban-board web page.
This has been included in the stylesheet.css file for the kanban-board web page. This makes the contenteditable div task cards become active with entering text. The "webkit-user-select; text" makes the mobile browser aware that there is a virtual keyboard input device on the browser. This seemed to be a problem on the apple iOS platform. When touching on the task cards to enter text on the touch keyboard, text would not be entered. Now works on all web browsers installed on iOS including chrome and safari.