Have enabled user session tokens for user login feature: continued.
With requiring the "crypto" dependency, this is used to encrypt and generate a session token that is assigned to an existing userID when they log in, thus assigning that userID to a sessionID with that token in the "user_login" db table. This section contains all of the Express API code that was written for this login functionality.

The "userID" and "token" objects have put inside the SELECT statement MySQL queries which grabs all of the rows from the 'user_login' table to see if an existing login token matches for the currently logged in user session/ ID. If the login session is not empty, the page will still render and will contain the token once that web page has been navigated to from another page. If there is no login session token that is valid, the user will be brought back to the login page as the "login.ejs" page will be rendered. These SELECT statement MySQL queries have been included in all of the app.get("/") API end points for each of the web pages that are rendered within the application apart from the "login" and "user-sign-up" web pages themselves.