Have included internal JavaScript on the Index.ejs page (localStorage: userID and token configuration).
Here the UserID and token data objects are being pulled from the API "index.js" file, so that they can be both retrieved and saved to local storage in the browser. The "userID" and "token" being pulled from the API takes place on lines 19-20. The current logged in the user with the UserID and token object is stored in the URL for each web page within the application. This is how the session is stored throughout the process.

For each of the eight web pages defined, there is a web page element variable and url variable which will grab the userID and token data object and apply it for that web page and it's URL. hrefs had to be defined for all web pages from and within the index.ejs web page so that the id names could be placed in them for linking the hrefs into the variable assignments.

localStorage.setItem and localStorage.getItem are used in the if statements at the top for setting the current userID & token data value and setting them once they have been saved to local storage from the index page upon logging in. This is so that the other web pages can "get" the current "userID" and "token" object from local storage using "getItem" as this would have been set upon logging in on the index page. The "getItem" configuration needs to be set on all of the other web pages that are rendered in the app.