localStorage.getItem configuration being set for "userID" & "token" object: home-table.hbs page
userID and token data objects for login system being grabbed from localstorage for the  home-table page. This is so that the logged in userID along with it's token session number is carried to the page URL.

href links being made to the other specific web pages that can be navigated from this web page only to carry across the userID and token number in the URL to those web pages.

A new XMLHttpRequest event listener has been made within the if statement for the "delete" class name that hyperlinks to the delete API action for a personnel record. This is so that the window for the web page can just reload. This had to be implemented as this was clashing with the login system.