localStorage.getItem configuration being set for "userID" & "token" object: view-personnel.hbs page
userID and token data objects for login system being grabbed from localstorage for the view-personnel page. This is so that the logged in userID along with it's token session number is carried to the page URL.

href links being made to the other specific web pages that can be navigated from this web page only to carry across the userID and token number in the URL to those web pages.