Have started to create toggle button on settings page for dark theme mode.
This has been made using pure CSS using two styling classes called "toggle-container" and "toggle-button". Using the necessary CSS properties, the button (which is the circle shape in it's container with a border-radius outline) will be given an event listener in JavaScript to toggle and activate dark mode with main CSS colours. The correct toggle-button and toggle-container colours will next be applied.