Extra colour variable names added to "global.css" file.
These were used for items within the personnel management table such as "h" HTML tags and styling classes that were already proposed by bootstrap. Some of these classes were overridden to provide more suitable colours for the UI so that there was good contrast. The same personnel management table colour variable names have been used in the "theme.css" file. This is so that the contrast in colours can be set in each variable to match the main dark colours.